WASH Survey Raw Data for Dhaka, Bangladesh

WASH survey raw data collected by WSUP in Dhaka, Bangladesh

surveyDataBGD

Format

A data frame with 217 variables and 1282 entries:

country

Country

ccode

Country Code

zone

Enumeration Area

type

Sampling Point Type

quadrat

Quadrat Number

hhid

Household ID

uniqueID

Unique ID

month

Month of Survey

year

Year of Survey

deployment

Deployment

enumerator

Enumerator

status

Status

responseCode

Response Code

draftDate

Drafted On

submissionDate

Submitted On

ipAddress

IP Address

surveyType

Survey type

psu

primary sampling unit (PSU) ID

psuTime

PSU ID Time Answered

latitude

PSU ID Location Answered latitude

longitude

PSU ID Location Answered longitude

accuracy

PSU ID Location Answered accuracy

altitude

PSU ID Location Answered altitude

area1

Thana

area1Other

Thana Other please specify specify

area1Time

Thana Time Answered

area2

Ward

slum

Slum

gender

Respondent gender

landOwnStatus

What is the respondent land ownership status

nWomen

Number of adult women aged 16 to 60

nMen

Number of adult men aged 16 to 60

nOldWomen

Number of older women aged _60

nOldMen

Number of older men aged _60

nGirls

Number of girls aged 4 15

nBoys

Number of boys aged 4 15

nInfants

Number of infants _3

nMobility

How many people with reduced mobility are there in this household

water1

Main source of drinking water

water2

Main source of drinking water Other please specify specify

water3

Is the service provision formal or informal

water4

Is the service provision formal or informal Don t Know

water5

Is main source of drinking water a WSUP supported facility

water6

Is main source of drinking water a WSUP supported facility Don t Know

water7

Typical number of hours per day that water is available

water8

Typical number of hours per day that water is available Don t Know

water9

Typical days per week that water is available

water10

Typical days per week that water is available Don t Know

water11

Is water typically available and sufficient throughout the whole year

water12

Is water typically available and sufficient throughout the whole year Comments

water13

Is water typically available and sufficient throughout the whole year Don t Know

water14

How long does it take to go there get water and come back

water15

How long does it take to go there get water and come back Don t Know

water16

Are you satisfied that the distance between your main water source and your home is acceptable

water17

Are you satisfied that the distance between your main water source and your home is acceptable Don t Know

water18

Do you consider that the queuing time at your main water source is acceptable

water19

Do you consider that the queuing time at your main water source is acceptable Not Applicable

water20

Do you consider that the queuing time at your main water source is acceptable Don t Know

water21

Which household member s usually collect the water

water22

Which household member s usually collect the water Other please specify specify

water23

Typical amount of water used by your household per person per day

water24

Typical amount of water used by your household per person per day Don t Know

water25

Do you consider the amount of water supplied from your primary water source to be sufficient to meet your household_s requirements each day

water26

Do you consider the amount of water supplied from your primary water source to be sufficient to meet your household_s requirements each day Don t Know

water27

Do you also use water from other water sources

water28

Do you also use water from other water sources Don t Know

water29

Which is your usual secondary water source

water30

Which is your usual secondary water source Other please specify specify

water31

Which is your usual secondary water source Don t Know

water32

Typical amount of money spent on water per week magnitude

water33

Typical amount of money spent on water per week units

water34

Typical amount of money spent on water per week Don t Know

water35

Are you satisfied with the price of your main water source

water36

Are you satisfied with the price of your main water source Don t Know

water37

Would you be willing to invest in improving your drinking water source

water38

Would you be willing to invest in improving your drinking water source Comments

water39

Would you be willing to invest in improving your drinking water source Don t Know

water40

Are you satisfied with the quality of your main water source

water41

Are you satisfied with the quality of your main water source Don t Know

water42

What would increase your satisfaction

water43

What would increase your satisfaction Other improvement please specify specify

water44

What would increase your satisfaction Don t Know

water45

Are you satisfied with the water pressure of your main water source

water46

Are you satisfied with the water pressure of your main water source Don t Know

water47

Who do you talk to if there is a problem with the water supply

water48

Who do you talk to if there is a problem with the water supply Other please specify specify

water49

Who do you talk to if there is a problem with the water supply Not Applicable

water50

Who do you talk to if there is a problem with the water supply Don t Know

water51

Previous source of drinking water

water52

Previous source of drinking water Other please specify specify

water53

Previous source of drinking water Don t Know

water54

Typical amount of money spent per week on the previous water source magnitude

water55

Typical amount of money spent per week on the previous water source units

water56

Typical amount of money spent per week on the previous water source Don t Know

water57

Why do you use the WSUP supported facility instead of the facility you used before

water58

Why do you use the WSUP supported facility instead of the facility you used before Other please specify specify

water59

Why do you use the WSUP supported facility instead of the facility you used before Don t Know

san1

Usage arrangement for household_s sanitation facility

san2

Usage arrangement for household_s sanitation facility Other please specify specify

san3

Usage arrangement for household_s sanitation facility Don t Know

san4

How many people share use of the communal shared toilet

san5

How many people share use of the communal shared toilet Don t Know

san6

What kind of toilet facility do members in your household usually use

san7

What kind of toilet facility do members in your household usually use Other please specify specify

san8

What kind of toilet facility do members in your household usually use Don t Know

san9

Where does your toilet flush to

san10

Where does your toilet flush to Other please specify specify

san11

Where does your toilet flush to Don t Know

san12

Is main sanitation facility a WSUP supported facility

san13

Is main sanitation facility a WSUP supported facility Not Applicable

san14

Is main sanitation facility a WSUP supported facility Don t Know

san15

Is water available at the sanitation facility

san16

Is water available at the sanitation facility Don t Know

san17

Are you satisfied with the status of the hand washing facilities at the sanitation facility

san18

Are you satisfied with the status of the hand washing facilities at the sanitation facility Don t Know

san19

What would increase your satisfaction 1

san20

What would increase your satisfaction Other please specify specify

san21

What would increase your satisfaction Don t Know 1

san22

Is there an electric light in the sanitation facility

san23

Is there an electric light in the sanitation facility Don t Know

san24

Does the sanitation facility have a lockable door

san25

Does the sanitation facility have a lockable door Don t Know

san26

Does the sanitation facility have a container for menstrual hygiene management

san27

Does the sanitation facility have a container for menstrual hygiene management Don t Know

san28

The last time your youngest child passed stools what was done to dispose of the stools

san29

The last time your youngest child passed stools what was done to dispose of the stools Not Applicable

san30

The last time your youngest child passed stools what was done to dispose of the stools Don t Know

san31

How long does it take to go there queue for the toilet if applicable and come back

san32

How long does it take to go there queue for the toilet if applicable and come back Don t Know

san33

Typical amount of money spent on the main sanitation facility per week magnitude

san34

Typical amount of money spent on the main sanitation facility per week units

san35

Typical amount of money spent on the main sanitation facility per week Don t Know

san36

Are you satisfied with the price of the sanitation facility

san37

Are you satisfied with the price of the sanitation facility Don t Know

san38

Did you make an initial investment in the sanitation facility

san39

Did you make an initial investment in the sanitation facility Don t Know

san40

How much was the initial investment magnitude

san41

How much was the initial investment units

san42

How much was the initial investment Don t Know

san43

Would you be willing to invest in improved sanitation facilities

san44

Would you be willing to invest in improved sanitation facilities Comments

san45

Would you be willing to invest in improved sanitation facilities Don t Know

san46

Are you satisfied with the cleanliness of the sanitation facilities

san47

Are you satisfied with the cleanliness of the sanitation facilities Don t Know

san48

What would increase your level of satisfaction

san49

What would increase your level of satisfaction Other please specify specify

san50

What would increase your level of satisfaction Don t Know

san51

Is everyone in the household able to get to the sanitation facility and use it

san52

Is everyone in the household able to get to the sanitation facility and use it Don t Know

san53

Who is not using them

san54

Who is not using them Other please specify specify

san55

Who is not using them Don t Know

san56

Why not

san57

Why not Other please specify specify

san58

Why not Don t Know

san59

Do all household members usually use the available sanitation facilities

san60

Do all household members usually use the available sanitation facilities Don t Know

san61

Who is not using them 1

san62

Who is not using them Other please specify specify 1

san63

Who is not using them Don t Know 1

san64

Why are they not using the facilities

san65

Why are they not using the facilities Don t Know

san66

Who do you talk to if there is a problem with the sanitation facility

san67

Who do you talk to if there is a problem with the sanitation facility Other please specify specify

san68

Who do you talk to if there is a problem with the sanitation facility Not Applicable

san69

Who do you talk to if there is a problem with the sanitation facility Don t Know

san70

Do the women and girls in your household feel safe using the sanitation facilities during the day and night

san71

Do the women and girls in your household feel safe using the sanitation facilities during the day and night Don t Know

san72

Why not 1

san73

Why not Don t Know 1

san74

Do the women and girls in the household feel comfortable using the sanitation facilities during menstruation

san75

Do the women and girls in the household feel comfortable using the sanitation facilities during menstruation Not Applicable

san76

Why not 2

san77

Why not Other please specify specify 1

san78

Why not Don t Know 2

san79

What materials are used by the women girls of your house during their menstruation cycle

san80

What materials are used by the women girls of your house during their menstruation cycle Other please specify specify

san81

How is the cloth washed

san82

How is the cloth washed Other please specify specify

san83

Where do you dispose of the used napkin pad

san84

Where do you dispose of the used napkin pad Other please specify specify

san85

When should you wash your hands

san86

When should you wash your hands Other please specify specify

san87

When should you wash your hands Don t Know

san88

Can you show me where you wash your hands

san89

Can you show me where you wash your hands Not Applicable

wash60

Can you show me where you store your drinking water

wash61

Can you show me where you store your drinking water Not Applicable

wash62

Can you show me where you store your drinking water Don t Know

san90

How often do you empty your pit septic tank

san91

How often do you empty your pit septic tank Other please specify specify

san92

How often do you empty your pit septic tank Not Applicable

san93

How often do you empty your pit septic tank Don t Know

san94

How much does it cost to empty magnitude

san95

How much does it cost to empty units

san96

How much does it cost to empty Not Applicable

san97

How much does it cost to empty Don t Know

san98

Who empties the pit

san99

Who empties the pit Other please specify specify

san100

Who empties the pit Not Applicable

san101

Who empties the pit Don t Know

san102

Who pays for the pit to be emptied

san103

Who pays for the pit to be emptied Other please specify specify

san104

Who pays for the pit to be emptied Not Applicable

san105

Who pays for the pit to be emptied Don t Know

san106

Are you satisfied with the pit emptying service

ppi1

How many household members are 12 years old or younger

ppi2

Do all household members ages 6 to 12 currently attend a school educational institution

ppi3

X In the past year did any household member ever do work for which he she was paid on a daily basis

ppi4

How many rooms does your household occupy excluding rooms used for business

ppi5

What is the main construction material of the walls of the main room

ppi6

Does the household own any televisions

ppi7

How many fans does the household own

ppi8

How many mobile phones does the household own

ppi9

Does the household own any bicycles motorcycle scooters or motor cars etc

ppi10

Does the household own or rent sharecrop mortgage in or out 51 or more decimals of cultivable agricultural land excluding uncultivable land and dwelling house homestead land

ppi11

Does the household own or rent sharecrop mortgage in or out 51 or more decimals of cultivable agricultural land excluding uncultivable land and dwelling house homestead land Time Answered